CAA-55, Feature, Reorganized directories, XML files and added configuration for swagger (#3)
