+++ .gitignore
@@ -1,6 +1,8 @@
 HELP.md
 target/
 .mvn/
+mvnw
+mvnw.cmd
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
 !**/src/test/**/target/
